+++ flake8_import_order/checker.py
@@ -1,4 +1,5 @@
 import ast
+import re
 
 import asttokens
 
@@ -8,6 +9,20 @@ from flake8_import_order import ImportVisitor
 from flake8_import_order.styles import lookup_entry_point
 
 DEFAULT_IMPORT_ORDER_STYLE = 'cryptography'
+NOQA_INLINE_REGEXP = re.compile(
+    # We're looking for items that look like this:
+    # ``# noqa``
+    # ``# noqa: E123``
+    # ``# noqa: E123,W451,F921``
+    # ``# NoQA: E123,W451,F921``
+    # ``# NOQA: E123,W451,F921``
+    # We do not care about the ``: `` that follows ``noqa``
+    # We do not care about the casing of ``noqa``
+    # We want a comma-separated list of errors
+    '# noqa(?:: (?P<codes>([A-Z][0-9]+(?:[,\s]+)?)+))?',
+    re.IGNORECASE
+)
+COMMA_SEPARATED_LIST_RE = re.compile(r'[,\s]')
 
 
 class ImportOrderChecker(object):
@@ -61,5 +76,26 @@ class ImportOrderChecker(object):
         style = style_cls(visitor.imports)
 
         for error in style.check():
-            if not pycodestyle.noqa(self.lines[error.lineno - 1]):
+            if not self.error_is_ignored(error):
                 yield self.error(error)
+
+    def error_is_ignored(self, error):
+        noqa_match = NOQA_INLINE_REGEXP.search(self.lines[error.lineno - 1])
+        if noqa_match is None:
+            return False
+
+        codes_str = noqa_match.groupdict()['codes']
+        if codes_str is None:
+            return True
+
+        codes = set(parse_comma_separated_list(codes_str))
+        if error.code in codes:
+            return True
+
+        return False
+
+
+def parse_comma_separated_list(value):
+    value = COMMA_SEPARATED_LIST_RE.split(value)
+    item_gen = (item.strip() for item in value)
+    return [item for item in item_gen if item]