+++ tests/test_cases/noqa.py
@@ -1,7 +1,9 @@
 # appnexus cryptography edited google pep8 smarkets
 import ast
-import sys
+
+import sys  # noqa: I202
 
 import os # noqa
 import unittest
 import X # noqa
+from . import B, C, A  # I201 # noqa: I101